+++ R/summary.emfrail.R
@@ -103,10 +103,18 @@ summary.emfrail <- function(object, ...) {
   ci_theta_low <- exp(with(fit, outer_m$minimum - 1.96 * sqrt(1/outer_m$hess)))
   ci_theta_high <- exp(with(fit, outer_m$minimum + 1.96 * sqrt(1/outer_m$hess)))
 
-  if(theta > 9000) {
+  # if theta was at the edge, then CI should show this....
+  if(theta > object$.control$opt_control$interval[2] - 0.1) {
     ci_theta_low <- theta
     ci_theta_high <- Inf
   }
+
+  # likelihood based confidence intervals
+  if(isTRUE(object$.control$ci_based_intervals)) {
+    ci_theta_low <- exp(object$outer_m$ltheta_low)
+    ci_theta_high <- exp(object$outer_m$ltheta_high)
+  }
+
   # for gamma and pvf theta is 1/variance
   # for stable the L.T. is exp(- c^(1 - theta / (theta + 1)))